feat: Reguliere Aanbod v0.1
Change functions to work for reguliere woningen.

The bot checks if it is possible to react on the woning.
After 3 successful aanmeldingen the browser closes.
The bot was written to work only with loting woningen, functions
have been renamed to start the conversion to a bot that also reacts on
the reguliere aanbod.
Also adds a test desktop notification using notify-send.