Fix headless loting woning & Urgent notifications
Since there were not loting woningen the website would redirect
back to reguliere woningen, the bot would then start reacting
on woningen since the page was not filtered for jongeren woningen
and the "Ik heb al gereageerd" notifications were not visible.

The notify-send command now sends a critical notification with no
timeout. This way the notfication has to be closed manually.